Move CachingResourceLoader to own module.
